fix: Change ObservedGeneration type from int to int64
Fixed compilation error where ObservedGeneration (int) was being compared
and assigned from Generation (int64). This aligns with Kubernetes best
practices where status.observedGeneration should match metadata.generation
type.

+++ api/v1alpha1/resultsexporter_types.go
@@ -215,7 +215,7 @@ type ResultsExporterStatus struct {
 
 	// Observed generation
 	// +optional
-	ObservedGeneration int `json:"observedGeneration,omitempty"`
+	ObservedGeneration int64 `json:"observedGeneration,omitempty"`
 }
 
 // ExportedResourceCounts tracks how many resources have been exported
